refactor: 커뮤니티 페이지 구현 에러 수정
정상 실행 / 이미지 추가 필요
