+++ BE/src/main/java/com/healthyreal/be/api/repository/community/PostRepository.java
@@ -1,4 +1,4 @@
-package com.healthyreal.be.api.repository;
+package com.healthyreal.be.api.repository.community;
 
 import com.healthyreal.be.api.entity.community.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
