+++ BE/src/main/java/com/healthyreal/be/api/service/CommunityService.java
@@ -3,8 +3,8 @@ package com.healthyreal.be.api.service;
 import com.healthyreal.be.api.entity.community.Post;
 import com.healthyreal.be.api.entity.community.Comment;
 import com.healthyreal.be.api.entity.user.Member;
-import com.healthyreal.be.api.repository.PostRepository;
-import com.healthyreal.be.api.repository.CommentRepository;
+import com.healthyreal.be.api.repository.community.PostRepository;
+import com.healthyreal.be.api.repository.community.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
